+++ ndrandomOutput_js.tcl
@@ -1,5 +1,5 @@
 #nDRandom
-#ndrandomOutput_tcl.tcl
+#ndrandomOutput_js.tcl
 #
 ##===================================================================
 #	Copyright (c) 2021 Yuji SODE <yuji.sode@gmail.com>
@@ -7,11 +7,11 @@
 #	This software is released under the MIT License.
 #	See LICENSE or http://opensource.org/licenses/mit-license.php
 ##===================================================================
-#module to output as Tcl file
+#module to output as JavaScript file
 #it requires namespace `::nDRandom`
 #
 #=== option ===
-# - option: tcl
+# - option: js
 #
 #=== generated script ===
 #the generated script is output in the current directory
@@ -63,7 +63,7 @@ proc ::nDRandom::output\$ {NAME} {
 	#generated script format: "${NAME}_info()" and "${NAME}_random(double)" in math functions
 	#
 		#___[filename to output]___
-		set FILENAME "${NAME}.tcl";
+		set FILENAME "${NAME}.js";
 		#------------
 	#
 	set i 0;
@@ -74,31 +74,28 @@ proc ::nDRandom::output\$ {NAME} {
 		#------------
 		#
 		#___[output timestamp and data]___
-		puts -nonewline $c "\#timestamp:[clock format [clock seconds]]";
-		puts -nonewline $c "\n\#[join $INFO \n\#]";
+		puts -nonewline $c "//timestamp:[clock format [clock seconds]]";
+		puts -nonewline $c "\n//[join $INFO \n//]";
 		#------------
 		#
 		#___[output script: info]___
-		puts -nonewline $c "\n\#it returns preset data";
-		#
-		puts -nonewline $c "\nproc ::tcl::mathfunc::${NAME}_info \{\} \{return \{[join $INFO \n]\}\;\}\;";
+		puts -nonewline $c "\n//it returns preset data";
+		puts -nonewline $c "\nMath.${NAME}_info=function\(\)\{return \`[join $INFO \n]\`\;\}\;";
 		#------------
 		#
 		#___[output script: random]___
-		puts -nonewline $c "\n\#it returns a random coordinates in n-th dimensions";
-		puts -nonewline $c "\n\# - \$double: a real number in range \(0.0, 1.0\)";
-		puts -nonewline $c "\n\#output coordinates: \{x1 dx1 x2 dx2 ... xn dxn\} where Xi = xi+c*dxi and c is a constant";
+		puts -nonewline $c "\n//it returns a random coordinates in n-th dimensions";
+		puts -nonewline $c "\n// - double: a real number in range \(0.0, 1.0\)";
+		puts -nonewline $c "\n//output coordinates: \[x1 dx1 x2 dx2 ... xn dxn\] where Xi = xi+c*dxi and c is a constant";
 		#
-		puts -nonewline $c "\nproc ::tcl::mathfunc::${NAME}_random \{double\} \{";
+		puts -nonewline $c "\nMath.${NAME}_random=function\(double\)\{";
 			#
-			puts -nonewline $c "\n\tset result \{\}\;";
-			puts -nonewline $c "set idx 0;";
-			puts -nonewline $c "set block 0;";
+			puts -nonewline $c "\n\tlet result=\[\],idx=0,block=0,u=0.0\;";
 			#
-			#$u = [0.0, 1.0], then $u => index
-			puts -nonewline $c "set u \[expr \{\$double<0?0.0:\$double\}\]\;";
-			puts -nonewline $c "set u \[expr \{\$double>1?1.0:\$double\}\]\;";
-			puts -nonewline $c "set u \[expr \{int\(floor\(\$u*$CELLS\)\)\}\]\;";
+			#u = [0.0, 1.0], then $u => index
+			puts -nonewline $c "u=double<0?0.0:double\;";
+			puts -nonewline $c "u=double>1?1.0:double\;";
+			puts -nonewline $c "u=Math.floor\(u*$CELLS\)\;";
 			#
 			#VARS: {min1 d1 d1 dx1 min2 d2 d1*d2 dx2 ... minj dj d1*d2*...*dj dxj}
 			foreach {e1 e2 e3 e4} $VARS {
@@ -111,24 +108,24 @@ proc ::nDRandom::output\$ {NAME} {
 				#
 				#($i<1)? {$idx = $u%$e2}: {$idx = $block%$e2}
 				if {$i<1} {
-					puts -nonewline $c "\n\tset idx \[expr \{\$u%$e2\}\]\;";
+					puts -nonewline $c "\n\tidx=u%$e2\;";
 				} else {
-					puts -nonewline $c "\n\tset idx \[expr \{\$block%$e2\}\]\;";
+					puts -nonewline $c "\n\tidx=block%$e2\;";
 				};
 				#
 				#$block = ceil($u/$e3) := ceil(idx/∏(di))
-				puts -nonewline $c "set block \[expr \{int\(ceil\(\$u/$e3\)\)\}\]\;";
+				puts -nonewline $c "block=Math.ceil\(u/$e3\)\;";
 				#
 				#$e1+$e4*$idx := min+dx*idx
-				puts -nonewline $c "lappend result \[expr \{$e1+$e4*\$idx\}\]\;";
+				puts -nonewline $c "result.push\($e1+$e4*idx\)\;";
 				#$e4 := dx
-				puts -nonewline $c "lappend result $e4\;";
+				puts -nonewline $c "result.push\($e4\)\;";
 				#
 				incr i 1;
 			};
 			#
-			puts -nonewline $c "\n\tunset u idx block\;";
-			puts -nonewline $c "return \$result\;";
+			puts -nonewline $c "\n\tu=idx=block=null\;";
+			puts -nonewline $c "return result\;";
 			#
 		puts -nonewline $c "\n\}\;";
 		#------------